Refactor appliance model: add SwitchState and move Appliance to device_appliance
