+++ custom_components/xcomfort_bridge/xcomfort/devices.py
@@ -10,8 +10,9 @@ from this single module while the implementation is split across multiple files.
 from .device_base import BridgeDevice
 
 # Re-export device classes
+from .device_appliance import Appliance
 from .device_climate import Heater, RcTouch
-from .device_light import Appliance, Light
+from .device_light import Light
 from .device_rocker import Rocker
 from .device_sensors import DoorSensor, DoorWindowSensor, WindowSensor
 from .device_shade import Shade
@@ -22,6 +23,7 @@ from .device_states import (
     RcTouchState,
     RockerSensorState,
     ShadeState,
+    SwitchState,
 )
 
 __all__ = [
@@ -43,5 +45,6 @@ __all__ = [
     "RockerSensorState",
     "Shade",
     "ShadeState",
+    "SwitchState",
     "WindowSensor",
 ]
